Float to Double in ampl-converter

+++ ampl-converter/src/test/java/com/powsybl/ampl/converter/AmplNetworkReaderTest.java
@@ -130,35 +130,35 @@ public class AmplNetworkReaderTest {
         Generator generator = network.getGenerator("GEN");
         VoltageLevel voltageLevel = generator.getTerminal().getVoltageLevel();
 
-        assertEquals(24.5, generator.getTargetV(), 0.0f);
-        assertEquals(607.0f, generator.getTargetP(), 0.0f);
-        assertTrue(Float.isNaN(generator.getTerminal().getP()));
-        assertEquals(301.0f, generator.getTargetQ(), 0.0f);
-        assertTrue(Float.isNaN(generator.getTerminal().getQ()));
+        assertEquals(24.5, generator.getTargetV(), 0.0);
+        assertEquals(607.0, generator.getTargetP(), 0.0);
+        assertTrue(Double.isNaN(generator.getTerminal().getP()));
+        assertEquals(301.0, generator.getTargetQ(), 0.0);
+        assertTrue(Double.isNaN(generator.getTerminal().getQ()));
 
         reader.readGenerators();
 
-        assertEquals(voltageLevel.getNominalV() * 1.01000f, generator.getTargetV(), 0.0f);
-        assertEquals(300.0f, generator.getTargetP(), 0.0f);
-        assertEquals(300.0f, generator.getTerminal().getP(), 0.0f);
-        assertEquals(150.0f, generator.getTargetQ(), 0.0f);
-        assertEquals(150.0f, generator.getTerminal().getQ(), 0.0f);
+        assertEquals(voltageLevel.getNominalV() * 1.01000, generator.getTargetV(), 0.0);
+        assertEquals(300.0, generator.getTargetP(), 0.0);
+        assertEquals(300.0, generator.getTerminal().getP(), 0.0);
+        assertEquals(150.0, generator.getTargetQ(), 0.0);
+        assertEquals(150.0, generator.getTerminal().getQ(), 0.0);
     }
 
     private void testLoads(Network network, AmplNetworkReader reader) throws IOException {
         Load load = network.getLoad("LOAD");
 
-        assertEquals(600.0f, load.getP0(), 0.0);
-        assertTrue(Float.isNaN(load.getTerminal().getP()));
-        assertEquals(200.0f, load.getQ0(), 0.0);
-        assertTrue(Float.isNaN(load.getTerminal().getQ()));
+        assertEquals(600.0, load.getP0(), 0.0);
+        assertTrue(Double.isNaN(load.getTerminal().getP()));
+        assertEquals(200.0, load.getQ0(), 0.0);
+        assertTrue(Double.isNaN(load.getTerminal().getQ()));
 
         reader.readLoads();
 
-        assertEquals(300.0f, load.getP0(), 0.0);
-        assertEquals(305.0f, load.getTerminal().getP(), 0.0);
-        assertEquals(150.0f, load.getQ0(), 0.0);
-        assertEquals(155.0f, load.getTerminal().getQ(), 0.0);
+        assertEquals(300.0, load.getP0(), 0.0);
+        assertEquals(305.0, load.getTerminal().getP(), 0.0);
+        assertEquals(150.0, load.getQ0(), 0.0);
+        assertEquals(155.0, load.getTerminal().getQ(), 0.0);
     }
 
     private void testRatioTapChanger(Network network, AmplNetworkReader reader) throws IOException {
@@ -208,8 +208,8 @@ public class AmplNetworkReaderTest {
         Optional<Bus> bx = network.getVoltageLevelStream().map(vl -> vl.getBusView().getBus("VLGEN_0")).filter(Objects::nonNull).findFirst();
         if (bx.isPresent()) {
             Bus b = bx.get();
-            assertTrue(Float.isNaN(b.getAngle()));
-            assertTrue(Float.isNaN(b.getV()));
+            assertTrue(Double.isNaN(b.getAngle()));
+            assertTrue(Double.isNaN(b.getV()));
         } else {
             fail("Bus not found");
         }
@@ -219,7 +219,7 @@ public class AmplNetworkReaderTest {
         Optional<Bus> bx2 = network.getVoltageLevelStream().map(vl -> vl.getBusView().getBus("VLGEN_0")).filter(Objects::nonNull).findFirst();
         if (bx2.isPresent()) {
             Bus b = bx2.get();
-            assertEquals((float) Math.toDegrees(2d), b.getAngle(), 0.0);
+            assertEquals(Math.toDegrees(2d), b.getAngle(), 0.0);
             assertEquals(180f * b.getVoltageLevel().getNominalV(), b.getV(), 0.0);
         } else {
             fail("Bus not found");
@@ -229,133 +229,133 @@ public class AmplNetworkReaderTest {
 
     private void testBranches(Network network, AmplNetworkReader reader) throws IOException {
         Line l = network.getLine("NHV1_NHV2_1");
-        assertTrue(Float.isNaN(l.getTerminal1().getP()));
-        assertTrue(Float.isNaN(l.getTerminal1().getQ()));
-        assertTrue(Float.isNaN(l.getTerminal2().getP()));
-        assertTrue(Float.isNaN(l.getTerminal2().getQ()));
+        assertTrue(Double.isNaN(l.getTerminal1().getP()));
+        assertTrue(Double.isNaN(l.getTerminal1().getQ()));
+        assertTrue(Double.isNaN(l.getTerminal2().getP()));
+        assertTrue(Double.isNaN(l.getTerminal2().getQ()));
 
         //NHV2_NLOAD
         TwoWindingsTransformer twt = network.getTwoWindingsTransformer("NHV2_NLOAD");
-        assertTrue(Float.isNaN(twt.getTerminal1().getP()));
-        assertTrue(Float.isNaN(twt.getTerminal1().getQ()));
-        assertTrue(Float.isNaN(twt.getTerminal2().getP()));
-        assertTrue(Float.isNaN(twt.getTerminal2().getQ()));
+        assertTrue(Double.isNaN(twt.getTerminal1().getP()));
+        assertTrue(Double.isNaN(twt.getTerminal1().getQ()));
+        assertTrue(Double.isNaN(twt.getTerminal2().getP()));
+        assertTrue(Double.isNaN(twt.getTerminal2().getQ()));
 
         reader.readBranches();
 
         Line l2 = network.getLine("NHV1_NHV2_1");
-        assertEquals(-100f, l2.getTerminal1().getP(), 0.0);
-        assertEquals(-110f, l2.getTerminal1().getQ(), 0.0);
-        assertEquals(-200f, l2.getTerminal2().getP(), 0.0);
-        assertEquals(-120f, l2.getTerminal2().getQ(), 0.0);
+        assertEquals(-100, l2.getTerminal1().getP(), 0.0);
+        assertEquals(-110, l2.getTerminal1().getQ(), 0.0);
+        assertEquals(-200, l2.getTerminal2().getP(), 0.0);
+        assertEquals(-120, l2.getTerminal2().getQ(), 0.0);
         TwoWindingsTransformer twt2 = network.getTwoWindingsTransformer("NHV2_NLOAD");
-        assertEquals(-100f, twt2.getTerminal1().getP(), 0.0);
-        assertEquals(-110f, twt2.getTerminal1().getQ(), 0.0);
-        assertEquals(-200f, twt2.getTerminal2().getP(), 0.0);
-        assertEquals(-120f, twt2.getTerminal2().getQ(), 0.0);
+        assertEquals(-100, twt2.getTerminal1().getP(), 0.0);
+        assertEquals(-110, twt2.getTerminal1().getQ(), 0.0);
+        assertEquals(-200, twt2.getTerminal2().getP(), 0.0);
+        assertEquals(-120, twt2.getTerminal2().getQ(), 0.0);
     }
 
     private void testThreeWindingTransBranches(Network network, AmplNetworkReader reader) throws IOException {
         ThreeWindingsTransformer twt = network.getThreeWindingsTransformer("3WT");
-        assertTrue(Float.isNaN(twt.getLeg1().getTerminal().getP()));
-        assertTrue(Float.isNaN(twt.getLeg1().getTerminal().getQ()));
-        assertTrue(Float.isNaN(twt.getLeg2().getTerminal().getP()));
-        assertTrue(Float.isNaN(twt.getLeg2().getTerminal().getQ()));
-        assertTrue(Float.isNaN(twt.getLeg3().getTerminal().getP()));
-        assertTrue(Float.isNaN(twt.getLeg3().getTerminal().getQ()));
+        assertTrue(Double.isNaN(twt.getLeg1().getTerminal().getP()));
+        assertTrue(Double.isNaN(twt.getLeg1().getTerminal().getQ()));
+        assertTrue(Double.isNaN(twt.getLeg2().getTerminal().getP()));
+        assertTrue(Double.isNaN(twt.getLeg2().getTerminal().getQ()));
+        assertTrue(Double.isNaN(twt.getLeg3().getTerminal().getP()));
+        assertTrue(Double.isNaN(twt.getLeg3().getTerminal().getQ()));
 
         reader.readBranches();
 
         ThreeWindingsTransformer twt2 = network.getThreeWindingsTransformer("3WT");
-        assertEquals(-101f, twt2.getLeg1().getTerminal().getP(), 0.0);
-        assertEquals(-111f, twt2.getLeg1().getTerminal().getQ(), 0.0);
-        assertEquals(-102f, twt2.getLeg2().getTerminal().getP(), 0.0);
-        assertEquals(-112f, twt2.getLeg2().getTerminal().getQ(), 0.0);
-        assertEquals(-103f, twt2.getLeg3().getTerminal().getP(), 0.0);
-        assertEquals(-113f, twt2.getLeg3().getTerminal().getQ(), 0.0);
+        assertEquals(-101, twt2.getLeg1().getTerminal().getP(), 0.0);
+        assertEquals(-111, twt2.getLeg1().getTerminal().getQ(), 0.0);
+        assertEquals(-102, twt2.getLeg2().getTerminal().getP(), 0.0);
+        assertEquals(-112, twt2.getLeg2().getTerminal().getQ(), 0.0);
+        assertEquals(-103, twt2.getLeg3().getTerminal().getP(), 0.0);
+        assertEquals(-113, twt2.getLeg3().getTerminal().getQ(), 0.0);
     }
 
     private void testDLBranches(Network network, AmplNetworkReader reader) throws IOException {
         DanglingLine dl = network.getDanglingLine("DL");
-        assertTrue(Float.isNaN(dl.getTerminal().getP()));
-        assertTrue(Float.isNaN(dl.getTerminal().getQ()));
+        assertTrue(Double.isNaN(dl.getTerminal().getP()));
+        assertTrue(Double.isNaN(dl.getTerminal().getQ()));
 
         reader.readBranches();
 
         DanglingLine dl2 = network.getDanglingLine("DL");
-        assertEquals(-100f, dl2.getTerminal().getP(), 0.0);
-        assertEquals(-110f, dl2.getTerminal().getQ(), 0.0);
+        assertEquals(-100, dl2.getTerminal().getP(), 0.0);
+        assertEquals(-110, dl2.getTerminal().getQ(), 0.0);
     }
 
     private void testHvdc(Network network, AmplNetworkReader reader) throws IOException {
         HvdcLine hl = network.getHvdcLine("L");
 
         assertEquals(HvdcLine.ConvertersMode.SIDE_1_INVERTER_SIDE_2_RECTIFIER, hl.getConvertersMode());
-        assertEquals(280f, hl.getActivePowerSetpoint(), 0.0);
+        assertEquals(280, hl.getActivePowerSetpoint(), 0.0);
 
         reader.readHvdcLines();
 
         HvdcLine hl2 = network.getHvdcLine("L");
         assertEquals(HvdcLine.ConvertersMode.SIDE_1_RECTIFIER_SIDE_2_INVERTER, hl2.getConvertersMode());
-        assertEquals(300f, hl2.getActivePowerSetpoint(), 0.0);
+        assertEquals(300, hl2.getActivePowerSetpoint(), 0.0);
     }
 
     private void testLcc(Network network, AmplNetworkReader reader) throws IOException {
         LccConverterStation lc = network.getLccConverterStation("C1");
-        assertEquals(100f, lc.getTerminal().getP(), 0.0);
-        assertEquals(50f, lc.getTerminal().getQ(), 0.0);
+        assertEquals(100, lc.getTerminal().getP(), 0.0);
+        assertEquals(50, lc.getTerminal().getQ(), 0.0);
 
         reader.readLccConverterStations();
 
-        assertEquals(200f, lc.getTerminal().getP(), 0.0);
-        assertEquals(75f, lc.getTerminal().getQ(), 0.0);
+        assertEquals(200, lc.getTerminal().getP(), 0.0);
+        assertEquals(75, lc.getTerminal().getQ(), 0.0);
 
     }
 
     private void testVsc(Network network, AmplNetworkReader reader) throws IOException {
         VscConverterStation vc = network.getVscConverterStation("C1");
-        assertEquals(100f, vc.getTerminal().getP(), 0.0);
-        assertEquals(50f, vc.getTerminal().getQ(), 0.0);
+        assertEquals(100, vc.getTerminal().getP(), 0.0);
+        assertEquals(50, vc.getTerminal().getQ(), 0.0);
         assertTrue(vc.isVoltageRegulatorOn());
-        assertTrue(Float.isNaN(vc.getReactivePowerSetpoint()));
-        assertEquals(405f, vc.getVoltageSetpoint(), 0.0);
+        assertTrue(Double.isNaN(vc.getReactivePowerSetpoint()));
+        assertEquals(405, vc.getVoltageSetpoint(), 0.0);
 
         reader.readVscConverterStations();
 
-        assertEquals(200f, vc.getTerminal().getP(), 0.0);
-        assertEquals(75f, vc.getTerminal().getQ(), 0.0);
+        assertEquals(200, vc.getTerminal().getP(), 0.0);
+        assertEquals(75, vc.getTerminal().getQ(), 0.0);
         assertTrue(vc.isVoltageRegulatorOn());
-        assertEquals(30f, vc.getReactivePowerSetpoint(), 0.0);
-        assertEquals(vc.getTerminal().getVoltageLevel().getNominalV() * 1.01000f, vc.getVoltageSetpoint(), 0.0);
+        assertEquals(30, vc.getReactivePowerSetpoint(), 0.0);
+        assertEquals(vc.getTerminal().getVoltageLevel().getNominalV() * 1.01000, vc.getVoltageSetpoint(), 0.0);
     }
 
     private void testShunts(Network network, AmplNetworkReader reader) throws IOException {
         ShuntCompensator sc = network.getShunt("C1_Filter1");
 
-        assertEquals(25.0f, sc.getTerminal().getQ(), 0.0);
+        assertEquals(25.0, sc.getTerminal().getQ(), 0.0);
 
         reader.readShunts();
 
         ShuntCompensator sc2 = network.getShunt("C1_Filter1");
 
-        assertEquals(30f, sc2.getTerminal().getQ(), 0.0);
+        assertEquals(30.0, sc2.getTerminal().getQ(), 0.0);
     }
 
     private void testSvc(Network network, AmplNetworkReader reader) throws IOException {
         StaticVarCompensator sv = network.getStaticVarCompensator("SVC2");
 
         assertEquals(RegulationMode.VOLTAGE, sv.getRegulationMode());
-        assertEquals(390f, sv.getVoltageSetPoint(), 0.0);
-        assertTrue(Float.isNaN(sv.getTerminal().getQ()));
+        assertEquals(390.0, sv.getVoltageSetPoint(), 0.0);
+        assertTrue(Double.isNaN(sv.getTerminal().getQ()));
 
         reader.readStaticVarcompensator();
 
         StaticVarCompensator sv2 = network.getStaticVarCompensator("SVC2");
         assertEquals(RegulationMode.REACTIVE_POWER, sv2.getRegulationMode());
-        assertEquals(400f, sv2.getVoltageSetPoint(), 0.0);
-        assertEquals(-30f, sv2.getReactivePowerSetPoint(), 0.0);
-        assertTrue(Float.isNaN(sv2.getTerminal().getP()));
-        assertEquals(30f, sv2.getTerminal().getQ(), 0.0);
+        assertEquals(400.0, sv2.getVoltageSetPoint(), 0.0);
+        assertEquals(-30.0, sv2.getReactivePowerSetPoint(), 0.0);
+        assertTrue(Double.isNaN(sv2.getTerminal().getP()));
+        assertEquals(30.0, sv2.getTerminal().getQ(), 0.0);
     }
 
 }
